Attackrange Displayed, need to get units on range (inside attackrange) Need to change movement to not allow to position inside another unit (unit block path) Signed-off-by: PHNvieira015 <pedro.hvieira@al.infnet.edu.br>

+++ Assets/Script/Stats/Equipment/Weapon.cs
@@ -11,7 +11,7 @@ public class Weapon : Equipment
     [Header("Offensive Stats")]
     public ItemClass weapon;
     public float WeaponDamageModifier;
-    public float WeaponRange;
+    public int WeaponRange;
 
     // Weapon and Armor Types
     public enum WeaponType { Sword, Bow, Staff, Axe }
